requirements.txt
django
pytesseract==0.3.13
pillow
tesseract
tesseract-ocr